requirements.txt
beautifulsoup4==4.12.3
nlpia2-wikipedia==1.5.18
numpy==2.2.1
pandas==2.2.3
plotly==5.24.1
requests==2.32.3
streamlit==1.41.1
yfinance==0.2.51
